Move Inspector component to folder
